app4/logging-service/requirements.txt
fastapi
uvicorn
pydantic
hazelcast-python-client==5.4.0
python-consul